move test cases for type detection to new file
